Move f1600x4 under Dilithium and restructure a bit
