+++ sign/dilithium/internal/common/aes.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"crypto/aes"
