+++ sign/dilithium/internal/common/field_test.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"flag"
