+++ sign/dilithium/internal/common/pack_test.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"math/rand"
