+++ sign/dilithium/internal/common/poly_test.go
@@ -1,4 +1,4 @@
-package internal
+package common
 
 import (
 	"math/rand"
