+++ sign/dilithium/mode3/internal/pack.go
@@ -1,7 +1,7 @@
 package internal
 
 import (
-	common "github.com/cloudflare/circl/sign/dilithium/internal"
+	"github.com/cloudflare/circl/sign/dilithium/internal/common"
 )
 
 // Writes p with norm less than or equal η into buf, which must be of
