+++ sign/dilithium/mode3/internal/pack_test.go
@@ -3,7 +3,7 @@ package internal
 import (
 	"testing"
 
-	common "github.com/cloudflare/circl/sign/dilithium/internal"
+	"github.com/cloudflare/circl/sign/dilithium/internal/common"
 )
 
 func TestPolyPackLeqEta(t *testing.T) {
