+++ sign/dilithium/mode4aes.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 	"io"
 
-	common "github.com/cloudflare/circl/sign/dilithium/internal"
+	"github.com/cloudflare/circl/sign/dilithium/internal/common"
 	"github.com/cloudflare/circl/sign/dilithium/mode4aes"
 )
