+++ sign/dilithium/templates/mode.templ.go
@@ -11,7 +11,7 @@ import (
 	"fmt"
 	"io"
 
-	common "github.com/cloudflare/circl/sign/dilithium/internal"
+	"github.com/cloudflare/circl/sign/dilithium/internal/common"
 	"github.com/cloudflare/circl/sign/dilithium/{{ .Pkg }}"
 )
