+++ sign/dilithium/templates/modePkg.templ.go
@@ -15,7 +15,7 @@ import (
 	"errors"
 	"io"
 
-	common "github.com/cloudflare/circl/sign/dilithium/internal"
+	"github.com/cloudflare/circl/sign/dilithium/internal/common"
 	"github.com/cloudflare/circl/sign/dilithium/{{ .Pkg }}/internal"
 )
